requirements.txt
fastapi
uvicorn[standard]
celery
httpx
beautifulsoup4==4.12.3
pydantic
pydantic-settings
python-dotenv
lxml
pika
psycopg[binary]==3.2.1
# Database
SQLAlchemy==2.0.30
alembic==1.13.1 